+++ test/model_test.dart
@@ -17,8 +17,9 @@ void main() {
     print(g);
     expect(g.toString(), '''
 digraph "" {
-  graph [style=rounded fontname="Calibri" rankdir=TB];
-  node [style=filled gradientangle=270 fontname="Calibri" fillcolor="lavender"];
+  graph [style=rounded fontname="Arial Black" fontsize=13 penwidth=2.6];
+  node [shape=rect style="filled,rounded" fontname=Arial fontsize=15 fillcolor=Lavender penwidth=1.3];
+  edge [penwidth=1.3];
   "a" [label="a"];
   "b" [label="b"];
   "c" [label="c"];
@@ -28,7 +29,7 @@ digraph "" {
 ''');
 
     // rankdir LR
-    g = Model(rankdir: 'LR');
+    g = Model();
     g.nodes['a'] = Node('a', 'a');
     g.nodes['b'] = Node('b', 'b');
     g.nodes['c'] = Node('c', 'c');
@@ -37,8 +38,9 @@ digraph "" {
     print(g);
     expect(g.toString(), '''
 digraph "" {
-  graph [style=rounded fontname="Calibri" rankdir=LR];
-  node [style=filled gradientangle=270 fontname="Calibri" fillcolor="lavender"];
+  graph [style=rounded fontname="Arial Black" fontsize=13 penwidth=2.6];
+  node [shape=rect style="filled,rounded" fontname=Arial fontsize=15 fillcolor=Lavender penwidth=1.3];
+  edge [penwidth=1.3];
   "a" [label="a"];
   "b" [label="b"];
   "c" [label="c"];
@@ -108,8 +110,9 @@ digraph "" {
     print(g);
     expect(g.toString(), '''
 digraph "" {
-  graph [style=rounded fontname="Calibri" rankdir=TB];
-  node [style=filled gradientangle=270 fontname="Calibri" fillcolor="lavender"];
+  graph [style=rounded fontname="Arial Black" fontsize=13 penwidth=2.6];
+  node [shape=rect style="filled,rounded" fontname=Arial fontsize=15 fillcolor=Lavender penwidth=1.3];
+  edge [penwidth=1.3];
   subgraph "cluster~zero" {
     label="zero";
     "a";
@@ -122,7 +125,7 @@ digraph "" {
 ''');
 
     // rankdirLR
-    g = Model(rankdir: 'LR');
+    g = Model();
     c0 = Subgraph('zero', 'zero');
     c0.nodes.add('a');
     c0.nodes.add('b');
@@ -133,8 +136,9 @@ digraph "" {
     print(g);
     expect(g.toString(), '''
 digraph "" {
-  graph [style=rounded fontname="Calibri" rankdir=LR];
-  node [style=filled gradientangle=270 fontname="Calibri" fillcolor="lavender"];
+  graph [style=rounded fontname="Arial Black" fontsize=13 penwidth=2.6];
+  node [shape=rect style="filled,rounded" fontname=Arial fontsize=15 fillcolor=Lavender penwidth=1.3];
+  edge [penwidth=1.3];
   subgraph "cluster~zero" {
     label="zero";
     "a";
@@ -161,8 +165,9 @@ digraph "" {
     print(g);
     expect(g.toString(), '''
 digraph "" {
-  graph [style=rounded fontname="Calibri" rankdir=TB];
-  node [style=filled gradientangle=270 fontname="Calibri" fillcolor="lavender"];
+  graph [style=rounded fontname="Arial Black" fontsize=13 penwidth=2.6];
+  node [shape=rect style="filled,rounded" fontname=Arial fontsize=15 fillcolor=Lavender penwidth=1.3];
+  edge [penwidth=1.3];
   subgraph "cluster~zero" {
     label="zero";
     "a";
@@ -255,7 +260,7 @@ digraph "" {
         Metrics(true, [], 10, 12, 1.1, [], 40, 12.3, 1.2, 1000, 250.0);
     print(metrics);
     expect(metrics.toString(),
-        '"metrics" [label=" isAcyclic: true \\l numNodes: 10  \\l numEdges: 12  \\l avgDegree: 1.1 \\l numOrphans: 0 \\l ccd: 40 \\l acd: 12.3 \\l nccd: 1.2 \\l totalSloc: 1000 \\l avgSloc: 250.0 \\l" shape=rect];');
+        '"metrics" [label=" isAcyclic: true \\l numNodes: 10  \\l numEdges: 12  \\l avgDegree: 1.1 \\l numOrphans: 0 \\l ccd: 40 \\l acd: 12.3 \\l nccd: 1.2 \\l totalSloc: 1000 \\l avgSloc: 250.0 \\l"];');
     print(jsonEncode(metrics));
     expect(jsonEncode(metrics),
         '{"isAcyclic":true,"firstCycle":[],"numNodes":10,"numEdges":12,"avgDegree":1.1,"orphans":[],"ccd":40,"acd":12.3,"nccd":1.2,"totalSloc":1000,"avgSloc":250.0}');